cryptos/requirements.txt
alembic==1.13.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
binance-connector==3.5.1
loguru==0.7.2
psycopg2-binary==2.9.9
aio-pika==9.3.1
